Fixed IS31FL3731 LEDs on keybow 2040
